# openapi/appconfiguration-3.0.1-custom.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Azure App Configuration",
    "description": "Azure App Configuration REST API",
    "version": "2024-09-01",
    "x-typespec-generated": [
      {
        "emitter": "@azure-tools/typespec-autorest"
      }
    ]
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "OAuth2Auth": ["https://azconfig.io/.default"]
    }
  ],
...
  "paths": {
    "/keys": {
      "get": {
        "summary": "Gets a list of keys.",
        "description": "Gets a list of keys.",
        "operationId": "GetKeys",
        "parameters": [
          {
            "$ref": "#/components/parameters/Azure.Core.Foundations.ApiVersionParameter"
          },
          {
            "name": "name",
            "in": "query",
            "description": "A filter for the name of the returned keys.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "After",
            "in": "query",
            "description": "Instructs the server to return elements that appear after the element referred\nto by the specified token.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "after"
            },
            "x-ms-client-name": "after"
          },
          {
            "name": "Sync-Token",
            "in": "header",
            "description": "Used to guarantee real-time consistency between requests.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "syncToken"
            },
            "x-ms-client-name": "syncToken"
          },
          {
            "name": "Accept-Datetime",
            "in": "header",
            "description": "Requests the server to respond with the state of the resource at the specified\ntime.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "acceptDatetime"
            },
            "x-ms-client-name": "acceptDatetime"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Sync-Token": {
                "description": "Used to guarantee real-time consistency between requests.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/vnd.microsoft.appconfig.keyset+json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResult"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyListResult"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/vnd.microsoft.appconfig.keyset+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-ms-examples": {
          "Gets a list of keys.": {
            "$ref": "./examples/GetKeys.json"
          }
        },
        "x-ms-pageable": {
          "nextLinkName": "@nextLink",
          "itemName": "items"
        }
      },
      "head": {
        "summary": "Requests the headers and status of the given resource.",
        "description": "Requests the headers and status of the given resource.",
        "operationId": "CheckKeys",
        "parameters": [
          {
            "$ref": "#/components/parameters/Azure.Core.Foundations.ApiVersionParameter"
          },
          {
            "name": "name",
            "in": "query",
            "description": "A filter for the name of the returned keys.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "After",
            "in": "query",
            "description": "Instructs the server to return elements that appear after the element referred\nto by the specified token.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "after"
            },
            "x-ms-client-name": "after"
          },
          {
            "name": "Sync-Token",
            "in": "header",
            "description": "Used to guarantee real-time consistency between requests.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "syncToken"
            },
            "x-ms-client-name": "syncToken"
          },
          {
            "name": "Accept-Datetime",
            "in": "header",
            "description": "Requests the server to respond with the state of the resource at the specified\ntime.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "acceptDatetime"
            },
            "x-ms-client-name": "acceptDatetime"
          },
          {
            "$ref": "#/components/parameters/Azure.Core.ClientRequestIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Sync-Token": {
                "description": "Used to guarantee real-time consistency between requests.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-ms-examples": {
          "Check keys": {
            "$ref": "./examples/CheckKeys.json"
          }
        }
      }
    },
...
    "/labels": {
      "get": {
        "summary": "Gets a list of labels.",
        "description": "Gets a list of labels.",
        "operationId": "GetLabels",
        "parameters": [
          {
            "$ref": "#/components/parameters/Azure.Core.Foundations.ApiVersionParameter"
          },
          {
            "name": "name",
            "in": "query",
            "description": "A filter for the name of the returned labels.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sync-Token",
            "in": "header",
            "description": "Used to guarantee real-time consistency between requests.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "syncToken"
            },
            "x-ms-client-name": "syncToken"
          },
          {
            "name": "After",
            "in": "query",
            "description": "Instructs the server to return elements that appear after the element referred\nto by the specified token.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "after"
            },
            "x-ms-client-name": "after"
          },
          {
            "name": "Accept-Datetime",
            "in": "header",
            "description": "Requests the server to respond with the state of the resource at the specified\ntime.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "acceptDatetime"
            },
            "x-ms-client-name": "acceptDatetime"
          },
          {
            "name": "$Select",
            "in": "query",
            "description": "Used to select what fields are present in the returned resource(s).",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["name"],
                "x-ms-enum": {
                  "name": "LabelFields",
                  "modelAsString": true,
                  "values": [
                    {
                      "name": "name",
                      "value": "name",
                      "description": "Name field."
                    }
                  ]
                }
              },
              "x-ms-client-name": "select"
            },
            "x-ms-client-name": "select"
          },
          {
            "$ref": "#/components/parameters/Azure.Core.ClientRequestIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Sync-Token": {
                "description": "Used to guarantee real-time consistency between requests.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/vnd.microsoft.appconfig.labelset+json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelListResult"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelListResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelListResult"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/vnd.microsoft.appconfig.labelset+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-ms-examples": {
          "Gets a list of labels.": {
            "$ref": "./examples/GetLabels.json"
          }
        },
        "x-ms-pageable": {
          "nextLinkName": "@nextLink",
          "itemName": "items"
        }
      },
      "head": {
        "summary": "Requests the headers and status of the given resource.",
        "description": "Requests the headers and status of the given resource.",
        "operationId": "CheckLabels",
        "parameters": [
          {
            "$ref": "#/components/parameters/Azure.Core.Foundations.ApiVersionParameter"
          },
          {
            "name": "name",
            "in": "query",
            "description": "A filter for the name of the returned labels.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Sync-Token",
            "in": "header",
            "description": "Used to guarantee real-time consistency between requests.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "syncToken"
            },
            "x-ms-client-name": "syncToken"
          },
          {
            "name": "After",
            "in": "query",
            "description": "Instructs the server to return elements that appear after the element referred\nto by the specified token.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "after"
            },
            "x-ms-client-name": "after"
          },
          {
            "name": "Accept-Datetime",
            "in": "header",
            "description": "Requests the server to respond with the state of the resource at the specified\ntime.",
            "schema": {
              "type": "string",
              "x-ms-client-name": "acceptDatetime"
            },
            "x-ms-client-name": "acceptDatetime"
          },
          {
            "name": "$Select",
            "in": "query",
            "description": "Used to select what fields are present in the returned resource(s).",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["name"],
                "x-ms-enum": {
                  "name": "LabelFields",
                  "modelAsString": true,
                  "values": [
                    {
                      "name": "name",
                      "value": "name",
                      "description": "Name field."
                    }
                  ]
                }
              },
              "x-ms-client-name": "select"
            },
            "x-ms-client-name": "select"
          },
          {
            "$ref": "#/components/parameters/Azure.Core.ClientRequestIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Sync-Token": {
                "description": "Used to guarantee real-time consistency between requests.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {}
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-ms-examples": {
          "Requests the headers and status of the given resource.": {
            "$ref": "./examples/CheckLabels.json"
          }
        }
      }
    },
...
    "/operations": {
      "get": {
        "summary": "Gets the state of a long running operation.",
        "description": "Gets the state of a long running operation.",
        "operationId": "GetOperationDetails",
        "parameters": [
          {
            "$ref": "#/components/parameters/Azure.Core.Foundations.ApiVersionParameter"
          },
          {
            "name": "snapshot",
            "in": "query",
            "description": "Snapshot identifier for the long running operation.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Azure.Core.ClientRequestIdHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationDetails"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-ms-examples": {
          "Gets the state of a long running operation.": {
            "$ref": "./examples/GetOperationStatus.json"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Azure.Core.Foundations.Error": {
        "required": ["code", "message"],
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "One of a server-defined set of error codes."
          },
          "message": {
            "type": "string",
            "description": "A human-readable representation of the error."
          },
          "target": {
            "type": "string",
            "description": "The target of the error."
          },
          "details": {
            "type": "array",
            "description": "An array of details about specific errors that led to this reported error.",
            "items": {
              "$ref": "#/components/schemas/Azure.Core.Foundations.Error"
            }
          },
          "innererror": {
            "$ref": "#/components/schemas/Azure.Core.Foundations.InnerError"
          }
        },
        "description": "The error object."
      },
      "Azure.Core.Foundations.InnerError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "One of a server-defined set of error codes."
          },
          "innererror": {
            "$ref": "#/components/schemas/Azure.Core.Foundations.InnerError"
          }
        },
        "description": "An object containing more specific information about the error. As per Azure REST API guidelines - https://aka.ms/AzureRestApiGuidelines#handling-errors."
      },
      "Azure.Core.Foundations.OperationState": {
        "type": "string",
        "description": "Enum describing allowed operation states.",
        "enum": ["NotStarted", "Running", "Succeeded", "Failed", "Canceled"],
        "x-ms-enum": {
          "name": "OperationState",
          "modelAsString": true,
          "values": [
            {
              "name": "NotStarted",
              "value": "NotStarted",
              "description": "The operation has not started."
            },
            {
              "name": "Running",
              "value": "Running",
              "description": "The operation is in progress."
            },
            {
              "name": "Succeeded",
              "value": "Succeeded",
              "description": "The operation has completed successfully."
            },
            {
              "name": "Failed",
              "value": "Failed",
              "description": "The operation has failed."
            },
            {
              "name": "Canceled",
              "value": "Canceled",
              "description": "The operation has been canceled by the user."
            }
          ]
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the error."
          },
          "title": {
            "type": "string",
            "description": "A brief summary of the error."
          },
          "name": {
            "type": "string",
            "description": "The name of the parameter that resulted in the error."
          },
          "detail": {
            "type": "string",
            "description": "A detailed description of the error."
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code that the error maps to.",
            "format": "int32"
          }
        },
        "description": "Azure App Configuration error object."
      },
      "Key": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the key.",
            "readOnly": true
          }
        },
        "description": "Keys serve as identifiers for key-values and are used to store and retrieve corresponding values."
      },
      "KeyListResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "The collection value.",
            "items": {
              "$ref": "#/components/schemas/Key"
            }
          },
          "@nextLink": {
            "type": "string",
            "description": "The URI that can be used to request the next set of paged results."
          }
        },
        "description": "The result of a list request."
      },
...
      "Label": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the label."
          }
        },
        "description": "Labels are used to group key-values."
      },
...
      "LabelListResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "The collection value.",
            "items": {
              "$ref": "#/components/schemas/Label"
            }
          },
          "@nextLink": {
            "type": "string",
            "description": "The URI that can be used to request the next set of paged results."
          }
        },
        "description": "The result of a list request."
      },
      "OperationDetails": {
        "required": ["id", "status"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique id of the operation."
          },
          "status": {
            "$ref": "#/components/schemas/Azure.Core.Foundations.OperationState"
          },
          "error": {
            "$ref": "#/components/schemas/Azure.Core.Foundations.Error"
          }
        },
        "description": "Details of a long running operation."
      },
...
    },
    "parameters": {
      "Azure.Core.ClientRequestIdHeader": {
        "name": "x-ms-client-request-id",
        "in": "header",
        "description": "An opaque, globally-unique, client-generated string identifier for the request.",
        "schema": {
          "type": "string",
          "format": "uuid",
          "x-ms-parameter-location": "method",
          "x-ms-client-name": "clientRequestId"
        },
        "x-ms-parameter-location": "method",
        "x-ms-client-name": "clientRequestId"
      },
      "Azure.Core.Foundations.ApiVersionParameter": {
        "name": "api-version",
        "in": "query",
        "description": "The API version to use for this operation.",
        "required": true,
        "schema": {
          "minLength": 1,
          "type": "string",
          "x-ms-parameter-location": "method",
          "x-ms-client-name": "apiVersion"
        },
        "x-ms-parameter-location": "method",
        "x-ms-client-name": "apiVersion"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "Connection String",
        "in": "header"
      },
      "OAuth2Auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/authorize",
            "scopes": {
              "https://azconfig.io/.default": ""
            }
          }
        }
      }
    }
  },
...
}